Moving pipeline conversions to same file as other renderer type conversions
